requirements.txt
fastapi==0.115.13
langchain_core==0.3.66
langchain_google_genai==2.1.5
langgraph==0.4.8
protobuf==6.31.1
pydantic==2.11.7
PyPDF2==3.0.1
python-dotenv==1.1.0
Requests==2.32.4
streamlit==1.44.1
uvicorn==0.34.3
pymongo==4.10.1
motor==3.7.0
